requirements.txt
# numpy
# pandas
# scikit-learn
# matplotlib
# seaborn
# joblib
# ipython
# ipywidgets
# tensorflow
# keras
# flask
# tf-keras
# flask
# pandas
# wikipedia
# faiss-cpu
# sentence-transformers

# # LangChain ecosystem
# langchain
# langchain-core
# langchain-community
# langchain-text-splitters

# # For Groq API
# langchain-groq
# groq

# # For async support (often required by LangChain)
# pydantic>=2.0


Flask
numpy
pandas
scikit-learn
joblib

faiss-cpu
sentence-transformers

langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-groq
groq

pydantic>=2.0
wikipedia
python-dotenv
